requirements.txt
psycopg2-binary
lxml
networkx
jinja2